fix(explorer): fix AI Fix button — disabled when no LLM, async createEffect, camelCase
Three bugs:
1. disabled={!aiStatus()} made the button permanently greyed out whenever
   fv_llm_status returned null (no LLM configured or DB issue). Button now
   always clickable; startAiFix() re-checks status and shows a clear inline
   error if no endpoint exists.
2. createEffect(async () => {...}) anti-pattern replaced with a plain async
   refreshAiStatus() function called from a sync createEffect — avoids
   SolidJS reactive tracking issues with async effects.
3. Fixed invoke param casing — callId now explicitly named.

+++ crates/minion-llm/src/streaming.rs
@@ -80,7 +80,7 @@ where
         // Process every complete newline-terminated SSE line
         while let Some(pos) = line_buf.find('\n') {
             let line = line_buf[..pos].to_string();
-            line_buf.drain(..=pos as usize);
+            line_buf.drain(..=pos);
 
             if let Some(event) = parse_sse_line(&line) {
                 match event {
